requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
python-dotenv
python-jose
passlib
bcrypt
pydantic
pydantic-settings
email-validator
groq
asyncio
alembic